Add solution which use math algorithm
The factorial of `number` is equal to the product of numbers from 1 to
`number`. The zero appears as a result of multiplying 2 and 5. Number
of 2 more then number of 5 => number of zeros is equal to the number
of fives in the decomposition `n!`on prime factors. This number is
equal to (n/5) + (n/5^2) + (n/5^3) + ... until the next summand
becomes 0.